test-requirements.txt
nose>=1.2.1,<2.0
PyHamcrest>=1.8.3,<2
pyflakes>=0.8.1,<0.9
tempman>=0.1.3,<0.2
